fix dependencies processing
